Correcting ambiguity in naming of sub-group datasets that contribute to the models.
A subgroup dataset that contains all studies of its category is now referred to simply by its category (eg dfm_moca, dfm_arthritis) rather than with an additional suffix of `_full`. This also includes the two sensitivity analyses datasets - dfm_hi and dfm_lo. Rationale: these are the default dataset+models used in the paper. It is now easy to verify that tables and figures are drawing on the relevant models/datasets

A subgroup dataset that contains only high-quality studies is no longer referred to by only its category (dfm_mmse) but referred to with an additional suffix (_high) eg dfm_mmse_high. Rationale: previously the manuscript aimed to report these sub-sub-groups where possible (so using terminology to treat as a default) - but as this was revised on advice some time ago, it is better to mark these out as 'special cases' as they are now deprecated for use only in the appendices.